Application_Full_Stack/projet/requirements.txt
starlette_exporter
sqlalchemy
pydantic
psycopg2
python-slugify
python-jose==3.3.0